Add pagination on PageSnapshot GetKeys and GetEntries.
After calling PageSnapshot.GetKeys(or Entries) the first results are
immediately returned. If all results don't fit in a mojo message, a
"token" is also returned, giving access to the remaining keys (or
entries).

Status is PARTIAL_RESULT if there are more results. Although adding
this status is not necessary (we can just check if |next_token| is
NULL) it helps Ledger users avoid the case where they read the
initial results, see an OK status and don't check the value of
|next_token|.

Change-Id: I26a1b01da7524f4baa53245e3fd9ad69ad63355c