cloud_provider: scope CloudProvider to a particular page.
Currently CloudProvider is scoped to a particular user (it assumes that
the Firebase instance it uses is already configured to point at a url
that is owned by the particular user), but one instance of CloudProvider
can be used to sync multiple ledgers owned by multiple apps of the same
user.

As we don't envision sharing any objects in the cloud between different
pages, let alone different ledgers, we can simplify CloudProvider by
scoping it to a particular page, and this is what is done in this patch.

This also maps better to the sync logic, as PageSync will be now able to
own CloudProvider, removing the concern about CloudProvider callbacks
outliving the PageSync instance.

Change-Id: I36d1e32621ffab5563b0cc668ecfd19bf874b77c