cloud_sync: Add CommitUpload.
CommitUpload uploads the storage objects associated with a commit and
the commit metadata itself.

Contract: Unsynced objects referenced by the commit are marked as synced
as they are uploaded. The commit notification is uploaded only once all
objects are uploaded. The entire commit is marked as synced once all
objects are uploaded and the notification is uploaded.

If the upload fails, the client can call Start() again to retry.

LE-42

Change-Id: I978526170e9effb1d83893fcfd7d618df6049f3d