Always create a page when receiving a Page request.
This is a workaround for LE-81 while LE-87 is still pending: by
always creating a page, we ensure cloud synchronization starts even
for pages created remotely. The proper fix would be to have storage
check the cloud for the page before returning NOT_FOUND on a GetPage
call.

LE-81 #done

Change-Id: I8015ec5899c22531d9bd51292c70b387286c553d