firebase: encode control characters in Firebase values.
Firebase documentation states that control characters 0-31 are not
allowed in DB keys. It turns out that they are not allowed in DB values
as well.

This patch makes us encode the control characters away in values as we
were already doing for keys, fixing 400 Bad Request problem seen when
syncing certain objects.

LE-111 #done

Change-Id: I58bddb4f429e17f9c02696a9014c4e019eeb7c5c