Add backoff and error message when reconnecting to network service.
This prints out a warning and uses a backoff when reconnecting to the
network service.

Side-benefit of this is that if we run `cloud_sync doctor` without
bootstrap, we will get some helpful hint instead of hanging:

```
Network service crashed or not configured in environment, trying to reconnect.
```

LE-77 #comment This handles network service-level errors.

Change-Id: I48801e5ffc174d09c33eac2a4997b8d8e7f585d5