cloud_sync: check if the given page is in the cloud upon GetPage()
This patch removes the hack that always created the requested page upon
GetPage(), even if it didn't exist neither locally nor in the cloud.

We now lookup the page in the cloud, and report success only if the page
is present in the cloud or the cloud is unreachable. In the latter case
we preserve the previous behavior (binding to an empty page) with a
warning.

Binding to an empty page when we can't get the content is a bit wonky,
but it can be useful, because when the connection does go back the the
sync kicks off and the data magically appears.

LE-87 #done

Change-Id: Ib331c3d533c96d6f113893f85ddd0f4788036324