Speak mx::vmos and mx::sockets in the GCS client.
This prepares the currently unused GCS client to power object storage in
sync.

The current API takes input and output as file paths to read/to write
to, but file paths are internal to storage. This patch changes the api
so that it speaks vmos and sockets, in line with what is needed for the
Cloud Provider API. We also add more precise error codes.

LE-125 #comment bazinga

Change-Id: I2d131ad6385b6d0fe46679547b8361f117967ae0