configuration: add gcs config.
This patch adds gcs_bucket to Ledger configuration and renames
firebase_prefix to user_prefix (to be used for both Firebase and GCS).

The --gcs_bucket is added here as optional and not documented. A
subsequent patch makes it required and documents the GCS configuration.

LE-125 #comment configuration: add gcs config.

Change-Id: I31212052ca193ae617eeba988557dfb40ba36993