gcs: switch to Firebase Storage from bare GCS.
This patch switches the GCS client to use the Firebase Storage API
endpoints rather than the bare GCS. This allows us in the future to use
one authentication token for both commit notifications and storage
objects.

Because Firebase Storage endpoints don't seem to handle slashes in
object names, we now dcheck that there are none. When GCS is plugged in,
this will be ensured by encoding in the cloud_sync layer.

LE-125 #comment gcs: switch to Firebase Storage from bare GCS.

Change-Id: Id0bb615a7ad4aa06a053bff9fb15d16d39ce1e5c