Stop the branch tracker also for PUT/DELETE out of a transaction.
Now that PUT and DELETE are not executed synchronously, it is necessary
to stop the tracker, otherwise the track might follow another branch
than the one the client thinks it is on.

Also refactor the branch tracker to take the new commit to follow in the
StopTransaction method instead of having a separate method for stopping
the transaction and following a new branch.

Change-Id: Ia1cc7537cd906c1e4bc529bc4761a3a71095b287