Add PageDelegate to re-organize PageImpl and BranchTracker.
Until now PageManager owned a set of BranchTrackers, each of which
owned a PageImpl, the implementation of the Page interface. This CL
adds PageDelegate: PageManager owns a set of PageDelegates and each
of these owns a BranchTracker and a PageImpl.

This change will allow blocking BranchTracker's and PageImpl's objects
deletion until all callbacks (e.g. from PageStorage) have terminated.

Change-Id: Ic2d39952e4ecf7e00200a3a9e3e25c8e16b14321