Implement the AUTOMATIC_WITH_FALLBACK merge strategy.
This implements the last merge strategy, which consists in automatically
merging conflicts when each branch modify different keys. If any key is
modified in both branches, merging is deferred to the client as with the
CUSTOM strategy.

To this effect, CustomMergeStrategy is refactored to share common logic
with AutoMergeStrategy.

LE-124 #done

Change-Id: Id175a554735a35a17ec0da66ca03136a2c287c37