cloud_sync: fix nesting of the data stored in Firebase.
We recently added one more, optional level of nesting in the Firebase DB
for --cloud_prefix. The structure of the db can't be optional though, as
we define indexes on it.

This patch fixes the number of nesting levels (by using a default value
for --cloud_prefix if not passed) and updates the documentation
accordingly.

LE-144 #done

Change-Id: I619a0682d3e912b6317837c412917b696d8ae519