benchmark: add a sync latency benchmark.
This patch adds a benchmark that measures sync latency between two
instances of the same Ledger.

The benchmark runs on a single devices, but emulates the cross-device
sync by creating two instances of the same user's Ledger and relies on
Cloud sync to synchronize data between them.

In the benchmark, we measure the time between writing a value to the
"alpha" instance and observing the value through the watch API on the
"beta" instance.

Change-Id: I50021543d0214e8d43e787167d7b2d35a91e15d0