cloud sync: remove the gcs_bucket config parameter
As we're using Firebase Storage, GCS bucket actually has to be *the* GCS
bucket associated with the Firebase project, and the id of the default
bucket is (as far as we know) reliably <firebase id>.appspot.com.

Removing the parameter until proven needed.

LE-183 #comment simplify the config before moving it over to fw

Change-Id: I47a56f0aab774cfa2c942a23a9ef9c303d172099