app: fix the common ancestor search running out of memory and add tests
Not sure why exactly we were seemingly running out of heap at 17MB, but
indeed we were hitting crashes in heap allocations. It turns out that
even though the function has an asynchronous hop to waiter finalizer,
the GetCommit calls managed by the waiter are actually synchronous. And
it appears that if the calls managed by waiter are synchronous, the
Finalize() call becomes synchronous too.

So we were really chaining up the whole search in a recursion stack,
one-level per commit.

This patch fixes this by adding a PostTask() in between. This is likely
not the best way and we may want to switch to a simple iterator using a
sync api, or to coroutines... In the meantime, this CL fixes the crash.

LE-187 #done

Change-Id: I3809d4e38c8b6c99d8cb9a7f21958c6a02a2c42c