cloud_sync: make the sync configuration per-user.
This patch drops the sync configuration from the global environment, and
instead keeps a separate user configuration per each ledger repository.
This lays ground for moving over the configuration to the framework,
which will provide this per-user config upon connecting to each repository.

Note that there is no functional changes intended in this patch - the
per-user configuration is still derived from the global config file, to
be changed in subsequent patches.

LE-183 #comment make the sync configuration per-user

Change-Id: I4d1e0a4a8ce320145936a7acad292c77a446219e