Scope `ledger_tool` commands to the "current" user.
(`ledger_tool` is a new name for `cloud_sync`, see
https://fuchsia-review.googlesource.com/c/27809/)

This patch ensures that the debugging commands `ledger_tool clean` and
`ledger_tool doctor` are scoped to the local user ID, so that
`ledger_tool clean` does not blow away the data on the entire cloud
provider, which can be shared by multiple users.

This is achieved in a hacky way by storing the id and the repo directory
of the most recent run of Ledger in a well-defined location. These files
are then read by `ledger_tool`. This is a stop-gap solution until
command-line utils can run in the user scope.

LE-183 #comment Scope `ledger_tool` commands to the "current" user.

Change-Id: I820554f9bfb64a807194bda960d73db4c92be7da