Ensure commit are uploaded in order.
On Commits, Page Storage was notifying observers only after calling its
callback. If the callback was commiting a new commit itself, the
notification were in the wrong order.

This solves this issue by ensuring that a reentrant commit is only send
to watcher by the most external call to AddCommits.

LE-196 #comment This is a partial fix.

Change-Id: I00917b802a29a436402b344a3fe4f9f1d6fb8044