cloud_sync: upload commits only when on single head
This patch pauses commit upload while there are more than one local
head, and resumes it when we're back on a single head. This avoids merge
storms where multiple devices simultaneously resolve the same conflict.

Note that this renders useless the merge policy of NONE - clients
selecting this policy won't be able to propagate their commits to the
cloud. As we don't have any users of this policy, perhaps we should
remove this option for now.

LE-205 #done

Change-Id: I6d2d06eec5988661958cbd3d2292b326a5f5a90a