Store metadata in the cloud and use it to detect skew.
This CL store minimal meta-information in the cloud so that the Ledger
can detect when there is a version skew between the data in the cloud
and the local data.

LE-200 #comment Ledger will now detect when there is a version skew.
LE-211 #done Ledger will retry the check version.

Change-Id: I37fd5a929657ab84c6eb185004472dac6ccebda3