Serialize GetSnapshot and Put calls.
This is necessary to ensure client can see modification to the page
without having to wait for the other operations callbacks.

Put calls where already serialized, but only after the value have
been saved in the storage. This means that if some calls where
serialized while the value was saved, these calls would appear out
of order.

LE-213 #done Fixed.

Change-Id: I8a45817dae0c4419353bacd18f7ca095b1b41c1a