cloud_sync: upload local commits in batches
This patch adds the ability to upload a consecutive batch of commits
atomically to the commit log in the cloud; and uses this ability to
ensure that any locally divergent commits are always uploaded only after
the divergence is resolved in one atomic batch along with the merge
commits.

This way, each batch upload adds one (and no more) heads when synced to
another device, limiting the possibility of a merge storm.

LE-209 #done
LE-101 #done

Change-Id: I2194c84c5fd84b30403c9d4da891013c0745fd71