benchmark: add a benchmark for sync convergence
Adds a benchmark that measures the time it takes to reconcile two
concurrent changes. In the benchmark there are two devices.  At each
step, both devices make a concurrent write, and we measure the time
until both changes are visible to both devices.

Compared with the existing sync benchmark, which excercises
one-directional sync, this benchmark tracks performance of the
bi-directional sync and concurrent change reconciliation.

Change-Id: I41d5f2c99964b914bfdf53145a494ee5bf44517b