cloud sync: add auth provider interface
This patch abstracts away the source of auth tokens, so that cloud sync
does not need to depend on details of the modular::auth::TokenProvider
interface.

In the immediate term, this allows us to complete auth integration
before the needed methods land in TokenProvider.

LE-216 #comment add auth provider interface

Change-Id: Ib0bde066f3d9606dc9b3d6fe197c924e90114193