Add a watcher for the synchronization state.
Adds a facility to track the synchronization state of a page and a full
Ledger repository. The user-global tracking would permit to display a
status icon in the notification bar, while the per-page tracking would
allow each app to know precisely the state of the pages it uses.

I also split download and upload state as clients might be interested in
only one of the two. If you think this is unnecessarily detailed or have
other suggestions, I am interested.

LE-222 #comment Defined API

Change-Id: Ife3a1bf2b8e0cfc723377fa12e67df545b8ea935