Order local commits to be synced by the generation, not timestamp.
This patch fixes the commits-out-of-order issue "once and for all".

The issue was that local commits were being returned in the order of
local timestamps, which does not guarantee a topological ordering for
merge commits, as they inherit the timestamp from their parent commits.

This patch changes storage so that unsynced commits are ordered by their
generation instead.

LE-248 #done

Change-Id: Id7fbd719080367bfde46af5e79205f6b1a62637c